authStore update

+++ app/Http/Controllers/AuthController.php
@@ -2,12 +2,21 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\User;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Validation\Rule;
+use Auth;
 use Illuminate\Http\Request;
+
+
 
 class AuthController extends Controller
 {
+
     public function login(Request $request)
     {
+
 
         $credentials = $request->validate([
             'email' => 'required|email',
@@ -23,19 +32,55 @@ class AuthController extends Controller
         $user = User::where('email', $request->email)->first();
         $token = $user->createToken('auth_token')->plainTextToken;
 
+
         return response()->json([
             'access_token' => $token,
             'token_type' => 'Bearer',
             'user' => $user
+        ], 200);
+    }
+
+    public function register(Request $request)
+    {
+        $validated = $request->validate([
+            'name' => ['required', 'string', 'max:255'],
+            'email' => [
+                'required',
+                'string',
+                'email',
+                'max:255',
+                Rule::unique(User::class),
+            ],
+            'password' => ['required', 'string', 'confirmed']
         ]);
+
+        if ($validated) {
+            $user = User::create([
+                'name' => $request['name'],
+                'email' => $request['email'],
+                'password' => Hash::make($request['password']),
+            ]);
+        }
+
+        return response()->json([
+            'message' => 'User created successfully',
+            'user' => $user
+        ], 201);
+
+
     }
 
     public function logout(Request $request)
     {
-        $request->user()->currentAccessToken()->delete();
 
-        return response()->json([
-            'message' => 'Logged out successfully'
-        ]);
+        if ($request->user()->currentAccessToken() instanceof \Laravel\Sanctum\PersonalAccessToken) {
+            $request->user()->currentAccessToken()->delete();
+        }
+        return response('', 204);
+    }
+
+    public function user(Request $request)
+    {
+        return $request->user();
     }
 }
